updated test vec

+++ Triangle.h
@@ -15,19 +15,19 @@ namespace rabbit
         Vec3 P1_P2; ///< P1 to P2 Vector
         Vec3 Normal; ///< Normal to the Triangle's plane
         Vec3 Centroid;
-        float Area;
+        double Area;
         bool IsDegenerate;
 
         bool IsPointWithinExtrudedTriangle(const Point& p)const;
 
-        std::pair<Point,float> ProjectPointOntoTrianglePlane(const Point& p) const;
+        std::pair<Point,double> ProjectPointOntoTrianglePlane(const Point& p) const;
 
-        std::pair<float,float> CalcBarycentricCoords(const Point& p)const;
+        std::pair<double,double> CalcBarycentricCoords(const Point& p)const;
 
-        std::pair<Point, float> CalcInPlaneDistanceToTriangle(const Point& p)const;
+        std::pair<Point, double> CalcInPlaneDistanceToTriangle(const Point& p)const;
 
-        Point CalcPointFromBarycentricCoords(const float u, const float v) const;
-        Point CalcPointFromBarycentricCoords(std::pair<float,float> coords)const;
+        Point CalcPointFromBarycentricCoords(const double u, const double v) const;
+        Point CalcPointFromBarycentricCoords(std::pair<double,double> coords)const;
 
         boost::optional<Point> CheckPointSegDist(const Vec3& origin,
                                                  const Vec3& seg,
